QuoteEntity model fix

+++ src/main/java/com/coinpaprika/apiclient/entity/QuoteEntity.kt
@@ -14,9 +14,9 @@ import kotlinx.android.parcel.Parcelize
 @Parcelize
 data class QuoteEntity(@SerializedName("price") val price: Double,
                        @SerializedName("volume_24h") val dailyVolume: Long,
-                       @SerializedName("volume_24h_change_24h") val dailyVolumeDailyChange: Long,
+                       @SerializedName("volume_24h_change_24h") val dailyVolumeDailyChange: Double,
                        @SerializedName("market_cap") val marketCap: Long,
-                       @SerializedName("market_cap_change_24h") val marketCapDailyChange: Long,
+                       @SerializedName("market_cap_change_24h") val marketCapDailyChange: Double,
                        @SerializedName("percent_change_1h") val percentChange1h: Double,
                        @SerializedName("percent_change_12h") val percentChange12h: Double,
                        @SerializedName("percent_change_24h") val percentChange24h: Double,
